Use `cider-tooling-eval`
Instead of using `cider-nrepl-request:eval`, which invokes the spinner
in the REPL, use `cider-tooling-eval`.

The real reason behind this change has nothing to do with the spinner
itself, but when the spinner is spinning, the cider refuses to evaluate
any new code. Which means, that while a buffer is flychecking, I can't
use the repl or re-compile code, using `cider-tooling-eval` let's the
checkers run in the background, only interacting with the user's
workflow when emacs is actually processing the results.